Complete friend_rescue
